Move just the Network enum to -chain, keep everything else in -network
