--- zebra-chain/src/network.rs
@@ -1,31 +1,0 @@
-//! Network-specific types.
-
-use crate::types::Magic;
-
-/// An enum describing the possible network choices.
-#[derive(Copy, Clone, Debug, Eq, PartialEq, Hash, Serialize, Deserialize)]
-pub enum Network {
-    /// The production mainnet.
-    Mainnet,
-    /// The testnet.
-    Testnet,
-}
-
-impl Network {
-    /// Get the magic value associated to this `Network`.
-    pub fn magic(self) -> Magic {
-        match self {
-            Network::Mainnet => magics::MAINNET,
-            Network::Testnet => magics::TESTNET,
-        }
-    }
-}
-
-/// Magic numbers used to identify different Zcash networks.
-pub mod magics {
-    use super::*;
-    /// The production mainnet.
-    pub const MAINNET: Magic = Magic([0x24, 0xe9, 0x27, 0x64]);
-    /// The testnet.
-    pub const TESTNET: Magic = Magic([0xfa, 0x1a, 0xf9, 0xbf]);
-}
